requirements.txt
datasets==2.16.1
faiss_cpu==1.7.4
joblib==1.5.2
langchain==1.2.10
liquid==0.0.1
neo4j==5.28.2
numpy==2.4.2
openai_harmony==0.0.8
pandas==3.0.0
pyserini==0.22.1
python-dotenv==1.2.1
rapidfuzz==3.14.3
regex==2025.9.18
scikit_learn==1.8.0
sentence_transformers==5.1.2
spacy==3.7.5
torch==2.8.0
tqdm==4.66.1
transformers==4.55.1
